+++ index.js
@@ -46,7 +46,6 @@ function run (mod) {
 
 module.exports = {
   direct: run(direct),
-  dirty:  run(direct), // Deprecated name
   sam:    run(sam),
   static: run(_static)
 }