+++ src/direct/zip.js
@@ -9,24 +9,23 @@ var zipdir = require('zip-dir')
  * @returns {Buffer} zipfile as a buffer
  */
 module.exports = function zipper (pathIn, callback) {
-
-  function winzip (pathToCode, callback) {
-    zipdir(pathToCode, callback)
-  }
-
-  function nixzip (pathToCode, callback) {
-    series([
-      function _read (callback) {
-        glob(join(pathToCode, '/*'), { dot: true }, callback)
-      },
-      function _zip (files, callback) {
-        zipit({
-          input: files,
-        }, callback)
-      },
-    ], callback)
-  }
-
   let zip = process.platform.startsWith('win') ? winzip : nixzip
   zip(pathIn, callback)
+}
+
+function winzip (pathToCode, callback) {
+  zipdir(pathToCode, callback)
+}
+
+function nixzip (pathToCode, callback) {
+  series([
+    function _read (callback) {
+      glob(join(pathToCode, '/*'), { dot: true }, callback)
+    },
+    function _zip (files, callback) {
+      zipit({
+        input: files,
+      }, callback)
+    },
+  ], callback)
 }