+++ src/sam/02-after/index.js
@@ -1,7 +1,6 @@
 let series = require('run-series')
 let appApex  = require('./00-get-app-apex')
 let staticDeploy = require('../../static')
-let patchRestAPI = require('./02-patch-rest-api')
 let maybeInvalidate = require('./03-maybe-invalidate')
 let deployWS = require('./04-deploy-ws')
 let warnings = require('./05-warnings')
@@ -9,7 +8,7 @@ let warnings = require('./05-warnings')
 module.exports = function after (params, callback) {
   let {
     inventory,
-    legacyAPI,
+    legacyCompat,
     pretty,
     production,
     prune,
@@ -22,11 +21,10 @@ module.exports = function after (params, callback) {
   } = params
 
   series([
-    appApex.bind({}, { inventory, legacyAPI, pretty, region, stackname, stage, ts, update }),
+    appApex.bind({}, { inventory, pretty, region, stackname, stage, ts, update }),
     staticDeploy.bind({}, { inventory, isFullDeploy: true, production, prune, region, stackname, verbose, update }),
-    patchRestAPI.bind({}, { legacyAPI, region, stackname, stage }),
     maybeInvalidate.bind({}, { inventory, region, stackname, stage }),
-    deployWS.bind({}, { inventory, region, stackname, stage }),
+    deployWS.bind({}, { inventory, legacyCompat, region, stackname, stage }),
     warnings.bind({}, { inventory, update })
   ], callback)
 }